requirements.txt
autopep8==2.0.2
evdev==1.6.1
MouseInfo==0.1.3
mss==7.0.1
packaging==23.0
Pillow==9.4.0
PyAutoGUI==0.9.53
pycodestyle==2.10.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.10
python-xlib==0.33
python3-xlib==0.15
pytweening==1.0.4
six==1.16.0
tomli==2.0.1
types-Pillow==9.4.0.17
types-PyAutoGUI==0.9.3.5
types-PyScreeze==0.1.0.7
